上传版本

+++ ZYYG/Libs/JHRefresh/JHRefreshConfig.m
@@ -43,7 +43,7 @@ NSString *const JHRefreshLastUpdateTimeFormat = @"yyyy-MM-dd HH:mm";
 {
     NSUserDefaults *userDefault = [NSUserDefaults standardUserDefaults];
     NSDictionary *refreshDic = [userDefault objectForKey:JHRefreshConfigKey];
-    NSString *time = [[refreshDic objectForKey:JHRefreshLastUpdateTimeKey] objectForKey:[NSString stringWithFormat:@"%@_%d",JHRefreshLastUpdateTimeKey,ID]];
+    NSString *time = [[refreshDic objectForKey:JHRefreshLastUpdateTimeKey] objectForKey:[NSString stringWithFormat:@"%@_%ld",JHRefreshLastUpdateTimeKey,(long)ID]];
     
     if(!time)
     {
@@ -66,7 +66,7 @@ NSString *const JHRefreshLastUpdateTimeFormat = @"yyyy-MM-dd HH:mm";
     {
         lastUpdateTimeDic = [NSMutableDictionary dictionary];
     }
-    [lastUpdateTimeDic setObject:[[self class] nowTimeString] forKey:[NSString stringWithFormat:@"%@_%d",JHRefreshLastUpdateTimeKey,ID]];
+    [lastUpdateTimeDic setObject:[[self class] nowTimeString] forKey:[NSString stringWithFormat:@"%@_%ld",JHRefreshLastUpdateTimeKey,(long)ID]];
     
     [userDefault setObject:refreshDic forKey:JHRefreshConfigKey];
     [userDefault synchronize];
